requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.31.0
azure-core==1.29.1
azure-storage-blob==12.17.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
Flask==2.3.2
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
pycparser==2.21
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
